wip: code compiling

+++ casper-private-auction-core/src/accounts.rs
@@ -1,5 +1,5 @@
 // pub const MARKETPLACE_COMMISSION: u16 = 25;
 // pub const MARKETPLACE_ACCOUNT: &str = "account-hash-7de52a3013f609faa38ae99af4350da6aa6b69bec0e4087ecae87c2b9486a265";
 
-pub const MARKETPLACE_COMMISSION: u16 = 25;
+pub const MARKETPLACE_COMMISSION: u32 = 25;
 pub const MARKETPLACE_ACCOUNT: &str = "account-hash-e1a2a648532b6333c66b4fe316bff945d4d51636f5bec52161331b9dd39d3122";
